src/requirements.txt
django
djangorestframework
pandas
